small fix for inverse kinematics
Regardings the state of the kinematics: In general, it works, but there
is a small offset of ~1cm. This offset is only translational, not
rotational and in joint-space, the error is in all joints except
AnklePitch (the AnklePitch is perfectly correct). The source of the
offset might be that the ankle roll can only be computed from the
hip_rp_intersect, not the hip_ry_intersect. But the hip_rp_intersect
depends on the yaw, so there is a circle. What a pity!